requirements.txt
rucio-clients
paramiko
